Ryan's changes from Idaho
Merging in the changes made at the Idaho regional.  Mainly,
marking what to comment out to disable autonomous, and changing
the camera brightness from 0 to 40.

(This commit keeps the SetInverted() initialization that was added by Cole
after Utah, but wasn't committed to github, and thus wasn't picked up by
Ryan in Idaho.  But that code was tested by Cole and Jason.)

The shooter piston needs to be changed to automatically retract after
use.  Drive team had a habit of forgetting to do that. #derp