Branch of Cole's and Zack's mixed changes, which break at runtime
Cole made changes to control of the intake and lower launch motors
either together, to load a ball for shooting, or separately to eject
the boulder into the low goal.

Zack introduced new code to configure autonomous from the SmartDashboard.

Both sets of changes are untested, and putting them together breaks the
robot, so that the first time you hit a button, the program crashes. #derp

-- JL